+++ client/src/app/(layout)/(home)/components/ModelText.tsx
@@ -1,9 +1,9 @@
 "use client";
 
-import { useModelStore } from "@/stores/useModelStore";
+import { useModelSelectStore } from "@/stores/useModelSelectStore";
 
 const ModelText = () => {
-  const { selectedModel } = useModelStore();
+  const selectedModel = useModelSelectStore((state) => state.selectedModel);
 
   return <h1 className="text-4xl font-medium text-foreground">{selectedModel?.model || "무엇을 도와드릴까요?"}</h1>;
 };
